+++ packages/client/src/store/CallState.ts
@@ -604,6 +604,13 @@ export class CallState {
   };
 
   /**
+   * Returns the comparator currently used to sort the participants.
+   */
+  getSortParticipantsBy = (): Comparator<StreamVideoParticipant> => {
+    return this.sortParticipantsBy;
+  };
+
+  /**
    * Gets the current value of an observable, or undefined if the observable has
    * not emitted a value yet.
    *