+++ sample-apps/react-native/dogfood/src/components/CallControls/BottomControls/index.tsx
@@ -3,24 +3,22 @@ import {
   ScreenShareToggleButton,
   ToggleAudioPublishingButton,
   ToggleVideoPublishingButton,
-  useCallStateHooks,
   useTheme,
 } from '@stream-io/video-react-native-sdk';
 import React, { useMemo, useState } from 'react';
-import { LayoutChangeEvent, StyleSheet, Text, View } from 'react-native';
-import { MoreActionsButton } from './MoreActionsButton';
+import { LayoutChangeEvent, StyleSheet, View } from 'react-native';
+import { MoreActionsButton } from '../MoreActionsButton';
 import { ParticipantsButton } from './ParticipantsButton';
 import { ChatButton } from './ChatButton';
 import { RecordCallButton } from './RecordCallButton';
-import { ClosedCaptions } from './ClosedCaptions';
+import { SubtitleContainer } from './SubtitleContainer';
 
 export type BottomControlsProps = Pick<
   CallContentProps,
   'supportedReactions'
 > & {
-  onChatOpenHandler?: () => void;
-  onParticipantInfoPress?: () => void;
-  unreadCountIndicator?: number;
+  onChatOpenHandler: (() => void) | null;
+  onParticipantInfoPress: () => void;
   toggleCallRecording: () => Promise<void>;
   isAwaitingResponse: boolean;
   isCallRecordingInProgress: boolean;
@@ -28,7 +26,6 @@ export type BottomControlsProps = Pick<
 
 export const BottomControls = ({
   onChatOpenHandler,
-  unreadCountIndicator,
   onParticipantInfoPress,
   toggleCallRecording,
   isAwaitingResponse,
@@ -61,50 +58,15 @@ export const BottomControls = ({
         </View>
         <View style={styles.right}>
           <ParticipantsButton onParticipantInfoPress={onParticipantInfoPress} />
-          <ChatButton
-            onPressHandler={onChatOpenHandler}
-            unreadBadgeCount={unreadCountIndicator}
-          />
+          {onChatOpenHandler && (
+            <ChatButton onPressHandler={onChatOpenHandler} />
+          )}
         </View>
       </View>
       {!!controlsContainerHeight && (
         <SubtitleContainer controlsContainerHeight={controlsContainerHeight} />
       )}
     </>
-  );
-};
-
-// speaking while muted and caption controls - aka subtitle on top of video
-const SubtitleContainer = ({
-  controlsContainerHeight,
-}: {
-  controlsContainerHeight: number;
-}) => {
-  const styles = useStyles();
-  const { useIsCallCaptioningInProgress, useMicrophoneState } =
-    useCallStateHooks();
-  const isCaptioningInProgress = useIsCallCaptioningInProgress();
-  const { isSpeakingWhileMuted } = useMicrophoneState();
-  const {
-    theme: {
-      variants: { insets },
-    },
-  } = useTheme();
-  if (!isCaptioningInProgress || !isSpeakingWhileMuted) {
-    return null;
-  }
-  return (
-    <View
-      style={[
-        styles.subtitleContainer,
-        { bottom: controlsContainerHeight + insets.bottom },
-      ]}
-    >
-      <ClosedCaptions />
-      <View style={styles.speakingLabelContainer}>
-        <Text style={styles.label}>{'You are muted. Unmute to speak.'}</Text>
-      </View>
-    </View>
   );
 };
 
@@ -120,18 +82,6 @@ const useStyles = () => {
           paddingHorizontal: theme.variants.spacingSizes.md,
           flexDirection: 'row',
           justifyContent: 'flex-start',
-        },
-        subtitleContainer: {
-          position: 'absolute',
-          left: 0,
-          right: 0,
-        },
-        speakingLabelContainer: {
-          backgroundColor: theme.colors.sheetPrimary,
-        },
-        label: {
-          textAlign: 'center',
-          color: theme.colors.textPrimary,
         },
         left: {
           flex: 2.5,